+++ src/App.tsx
@@ -7,7 +7,7 @@ import {
 
 import BaseLayout from "components/BaseLayout";
 
-import { useLoadCurrentUser } from "./redux/auth";
+import { useLoadCurrentUser, usePermissions } from "./redux/auth";
 import LoginPage from "./pages/LogIn";
 import { PeoplePage, PersonPage } from "./pages/People";
 import { AllGearPage, GearItemPage } from "./pages/Gear";
@@ -22,10 +22,12 @@ import { ApproveDeskCreditPage } from "pages/OfficeHours/ApproveDeskCreditPage";
 import { OfficeHoursHistory } from "pages/OfficeHours/OfficeHoursHistory";
 import { MyOfficeHoursHistory } from "pages/OfficeHours/MyOfficeHoursHistory";
 import { ApprovalsPage } from "pages/Approvals";
+import { AddNewApproval } from "pages/Approvals/AddNewApproval";
 
 function App() {
   const { loggedIn, isLoading } = useLoadCurrentUser();
   const { pathname, search } = window.location;
+  const { isApprover } = usePermissions();
   return (
     <Router basename={process.env.PUBLIC_URL}>
       <BaseLayout>
@@ -74,6 +76,11 @@ function App() {
           <Route exact path="/approvals">
             <ApprovalsPage />
           </Route>
+          {isApprover && (
+            <Route exact path="/add-approval">
+              <AddNewApproval />
+            </Route>
+          )}
           {loggedIn && (
             <Route exact path="/request-desk-credit">
               <RequestDeskCreditPage />
